feat!: implement new Parser-js interface + migrate to TypeScript
