+++ .eslintignore
@@ -1,9 +1,8 @@
 node_modules
-.idea/*
 .vscode
 .DS_Store
 /docs
 /coverage
+/lib
 /esm
 /cjs
-*.tgz
