+++ README.md
@@ -2,6 +2,13 @@
 
 An AsyncAPI schema parser for Avro 1.x schemas.
 
+> **Note**
+> Version >= `2.0.0` of this package is only supported by the `@asyncapi/parser` version >= `2.0.0`.
+
+<!-- toc is generated with GitHub Actions do not remove toc markers -->
+
+<!-- toc -->
+
 ## Installation
 
 ```
@@ -11,10 +18,11 @@ npm install @asyncapi/avro-schema-parser
 ## Usage
 
 ```js
-const parser = require('@asyncapi/parser')
-const avroSchemaParser = require('@asyncapi/avro-schema-parser')
+import { Parser } from '@asyncapi/parser';
+import { AvroSchemaParser } from '@asyncapi/avro-schema-parser';
 
-parser.registerSchemaParser(avroSchemaParser);
+const parser = new Parser();
+parser.registerSchemaParser(AvroSchemaParser()); 
 
 const asyncapiWithAvro = `
 asyncapi: 2.0.0
@@ -40,16 +48,17 @@ channels:
               type: int
 `
 
-await parser.parse(asyncapiWithAvro)
+const { document } = await parser.parse(asyncapiWithAvro);
 ```
 
 ### Usage with remote references
 
 ```js
-const parser = require('@asyncapi/parser')
-const avroSchemaParser = require('@asyncapi/avro-schema-parser')
+import { Parser } from '@asyncapi/parser';
+import { AvroSchemaParser } from '@asyncapi/avro-schema-parser';
 
-parser.registerSchemaParser(avroSchemaParser)
+const parser = new Parser();
+parser.registerSchemaParser(AvroSchemaParser()); 
 
 const asyncapiWithAvro = `
 asyncapi: 2.0.0
@@ -65,16 +74,17 @@ channels:
           $ref: 'https://schemas.example.com/user'
 `
 
-await parser.parse(asyncapiWithAvro)
+const { document } = await parser.parse(asyncapiWithAvro);
 ```
 
 ### Usage with local references
 
 ```js
-const parser = require('@asyncapi/parser')
-const avroSchemaParser = require('@asyncapi/avro-schema-parser')
+import { Parser } from '@asyncapi/parser';
+import { AvroSchemaParser } from '@asyncapi/avro-schema-parser';
 
-parser.registerSchemaParser(avroSchemaParser)
+const parser = new Parser();
+parser.registerSchemaParser(AvroSchemaParser()); 
 
 const asyncapiWithAvro = `
 asyncapi: 2.0.0
@@ -90,7 +100,7 @@ channels:
           $ref: 'local/path/to/file/user'
 `
 
-await parser.parse(asyncapiWithAvro)
+const { document } = await parser.parse(asyncapiWithAvro);
 ```
 
 ### Usage with Confluent Schema Registry
@@ -106,10 +116,11 @@ await parser.parse(asyncapiWithAvro)
 #### Use them on your AsyncAPI document
 
 ```js
-const parser = require('@asyncapi/parser')
-const avroSchemaParser = require('@asyncapi/avro-schema-parser')
+import { Parser } from '@asyncapi/parser';
+import { AvroSchemaParser } from '@asyncapi/avro-schema-parser';
 
-parser.registerSchemaParser(avroSchemaParser)
+const parser = new Parser();
+parser.registerSchemaParser(AvroSchemaParser()); 
 
 const asyncapiWithAvro = `
 asyncapi: 2.0.0
@@ -125,21 +136,21 @@ channels:
           $ref: 'https://LY422RBU2HN6JQ5T:+f8wz9a0iM06AX7xfwbzSM9YPw/JIkr22Cvl5EKT5Hb1d/nz5nOpbXV/vZC+Iz5c@example.europe-west3.gcp.confluent.cloud/subjects/test/versions/1/schema'
 `
 
-await parser.parse(asyncapiWithAvro)
+const { document } = await parser.parse(asyncapiWithAvro);
 ```
 
 ## Features
 
 ### Validation of Avro schemas
 
-Avro schemas included in parsed AsyncAPI documents are validated using [avsc](https://www.npmjs.com/package/avsc). Invalid Avro schemas will cause the `parse` function to reject the promise, with an error that explains the problem.
+Avro schemas included in parsed AsyncAPI documents are validated using [avsc](https://www.npmjs.com/package/avsc). Invalid Avro schemas will cause the `parse` function to return `diagnostics` with all the validation errors.
 
 ```js
-const assert = require('assert');
-const parser = require('@asyncapi/parser');
-const avroSchemaParser = require('./index');
+import { Parser } from '@asyncapi/parser';
+import { AvroSchemaParser } from '@asyncapi/avro-schema-parser';
 
-parser.registerSchemaParser(avroSchemaParser);
+const parser = new Parser();
+parser.registerSchemaParser(AvroSchemaParser()); 
 
 const asyncapiWithInvalidAvro = `
 asyncapi: 2.0.0
@@ -155,11 +166,8 @@ channels:
           type: notAValidAvroType
 `;
 
-parser.parse(asyncapiWithInvalidAvro)
-  .catch(err => {
-    assert.strictEqual(err.message,
-      'unknown type: "notAValidAvroType"');
-  });
+const { document, diagnostics } = await parser.parse(doc);
+console.log(diagnostics);
 ```
 
 ### Support of required attributes
