+++ package.json
@@ -1,12 +1,16 @@
 {
   "name": "@asyncapi/avro-schema-parser",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "An AsyncAPI schema parser for Avro 1.x schemas.",
-  "main": "index.js",
+  "main": "cjs/index.js",
   "scripts": {
-    "test": "jest",
+    "build": "npm run build:esm && npm run build:cjs",
+    "build:esm": "tsc",
+    "build:cjs": "tsc --project ./tsconfig.cjs.json",
+    "test": "cross-env CI=true jest --coverage",
+    "lint": "eslint --max-warnings 0 --config .eslintrc .",
+    "lint:fix": "eslint --max-warnings 0 --config .eslintrc . --fix",
     "release": "semantic-release",
-    "lint": "eslint --max-warnings 0 --config .eslintrc.yaml .",
     "generate:assets": "echo 'No additional assets need to be generated at the moment'",
     "bump:version": "npm --no-git-tag-version --allow-same-version version $VERSION"
   },
@@ -23,28 +27,64 @@
     "schema",
     "parser"
   ],
-  "author": "Fran Mendez (fmvilas.com)",
+  "files": [
+    "/esm",
+    "/cjs",
+    "LICENSE",
+    "README.md"
+  ],
   "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/asyncapi/avro-schema-parser/issues"
   },
-  "homepage": "https://github.com/asyncapi/avro-schema-parser#readme",
+  "homepage": "https://github.com/asyncapi/avro-schema-parser",
+  "dependencies": {
+    "avsc": "^5.7.5",
+    "fetch-mock-jest": "^1.5.1",
+    "js-yaml": "^4.1.0",
+    "jsonpath-plus": "^7.2.0",
+    "lodash": "^4.17.21",
+    "node-fetch": "2.6.7",
+    "ramldt2jsonschema": "^1.2.3",
+    "webapi-parser": "^0.5.0"
+  },
   "devDependencies": {
-    "@asyncapi/parser": "^1.17.0",
+    "@asyncapi/parser": "^2.0.0-next-major.7",
+    "@jest/types": "^29.2.1",
     "@semantic-release/commit-analyzer": "^8.0.1",
-    "@semantic-release/github": "^7.0.4",
+    "@semantic-release/github": "^7.2.3",
     "@semantic-release/npm": "^7.1.1",
     "@semantic-release/release-notes-generator": "^9.0.1",
+    "@swc/jest": "^0.2.23",
+    "@types/jest": "^29.2.0",
+    "@types/json-schema": "^7.0.11",
     "conventional-changelog-conventionalcommits": "^4.5.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-jest": "^23.18.0",
-    "eslint-plugin-sonarjs": "^0.5.0",
-    "jest": "^26.1.0",
-    "semantic-release": "^17.4.2"
+    "cross-env": "^7.0.3",
+    "eslint": "^8.23.0",
+    "eslint-plugin-github": "^4.3.7",
+    "eslint-plugin-security": "^1.5.0",
+    "eslint-plugin-sonarjs": "^0.15.0",
+    "jest": "^29.0.2",
+    "semantic-release": "^17.4.2",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.2"
   },
   "release": {
     "branches": [
-      "master"
+      "master",
+      {
+        "name": "next-spec",
+        "prerelease": true
+      },
+      {
+        "name": "next-major-spec",
+        "prerelease": true
+      },
+      {
+        "name": "next-major",
+        "prerelease": true
+      }
     ],
     "plugins": [
       [
@@ -60,10 +100,12 @@
         }
       ],
       "@semantic-release/npm",
-      "@semantic-release/github"
+      [
+        "@semantic-release/github",
+        {
+          "assets": []
+        }
+      ]
     ]
-  },
-  "dependencies": {
-    "avsc": "^5.7.3"
   }
 }
