+++ src/index.ts
@@ -1,4 +1,63 @@
-const avsc = require('avsc');
+import avsc from 'avsc';
+
+import type { JSONSchema7TypeName } from 'json-schema';
+import type { Schema } from 'avsc';
+
+import type { AsyncAPISchema, SchemaParser, ParseSchemaInput, ValidateSchemaInput, SchemaValidateResult, SpecTypesV2 } from '@asyncapi/parser';
+
+type AvroSchema = Schema & { [key: string]: any };
+
+export function AvroSchemaParser(): SchemaParser {
+  return {
+    validate,
+    parse,
+    getMimeTypes,
+  };
+}
+
+async function validate(input: ValidateSchemaInput<unknown, unknown>): Promise<SchemaValidateResult[]> {
+  const result: SchemaValidateResult[] = [];
+  
+  try {
+    validateAvroSchema(input.data as AvroSchema);  
+  } catch (error) {
+    if (error instanceof Error) {
+      result.push({
+        message: error.message,
+        path: input.path, // avsc doesn't throw errors with meaningful paths
+      });
+    }
+  }
+  
+  return result;
+}
+
+async function parse(input: ParseSchemaInput<unknown, unknown>): Promise<AsyncAPISchema> {
+  const asyncAPISchema = await avroToJsonSchema(input.data as AvroSchema);
+
+  // TODO: Should the following modifications to the message object be done in the caller and for all parsers rather than here?
+  // remove that function when https://github.com/asyncapi/spec/issues/622 will be introduced in AsyncAPI spec
+  const message = (input.meta as any).message;
+  const key = message?.bindings?.kafka?.key;
+  if (key) {
+    const bindingsTransformed = await avroToJsonSchema(key);
+    message['x-parser-original-bindings-kafka-key'] = key;
+    message.bindings.kafka.key = bindingsTransformed;
+  }
+
+  return asyncAPISchema;
+}
+
+function getMimeTypes() {
+  return [
+    'application/vnd.apache.avro;version=1.9.0',
+    'application/vnd.apache.avro+json;version=1.9.0',
+    'application/vnd.apache.avro+yaml;version=1.9.0',
+    'application/vnd.apache.avro;version=1.8.2',
+    'application/vnd.apache.avro+json;version=1.8.2',
+    'application/vnd.apache.avro+yaml;version=1.8.2'
+  ];
+}
 
 const BYTES_PATTERN = '^[\u0000-\u00ff]*$';
 const INT_MIN = Math.pow(-2, 31);
@@ -6,7 +65,7 @@ const INT_MAX = Math.pow(2, 31) - 1;
 const LONG_MIN = Math.pow(-2, 63);
 const LONG_MAX = Math.pow(2, 63) - 1;
 
-const typeMappings = {
+const typeMappings: Record<string, JSONSchema7TypeName> = {
   null: 'null',
   boolean: 'boolean',
   int: 'integer',
@@ -23,7 +82,7 @@ const typeMappings = {
   uuid: 'string',
 };
 
-const commonAttributesMapping = (avroDefinition, jsonSchema, recordCache) => {
+function commonAttributesMapping(avroDefinition: AvroSchema, jsonSchema: SpecTypesV2.AsyncAPISchemaDefinition, recordCache: { [key:string]: AsyncAPISchema }): void {
   if (avroDefinition.doc) jsonSchema.description = avroDefinition.doc;
   if (avroDefinition.default !== undefined) jsonSchema.default = avroDefinition.default;
 
@@ -31,9 +90,9 @@ const commonAttributesMapping = (avroDefinition, jsonSchema, recordCache) => {
   if (fullyQualifiedName !== undefined && recordCache[fullyQualifiedName]) {
     jsonSchema['x-parser-schema-id'] = fullyQualifiedName;
   }
-};
+}
 
-function getFullyQualifiedName(avroDefinition) {
+function getFullyQualifiedName(avroDefinition: AvroSchema) {
   let name;
 
   if (avroDefinition.name) {
@@ -53,7 +112,7 @@ function getFullyQualifiedName(avroDefinition) {
  * @param parentJsonSchema the parent json schema which contains the required property to enrich
  * @param haveDefaultValue we assure that a required field does not have a default value
  */
-const requiredAttributesMapping = (fieldDefinition, parentJsonSchema, haveDefaultValue) => {
+function requiredAttributesMapping(fieldDefinition: any, parentJsonSchema: SpecTypesV2.AsyncAPISchemaDefinition, haveDefaultValue: boolean): void {
   const isUnionWithNull = Array.isArray(fieldDefinition.type) && fieldDefinition.type.includes('null');
 
   // we assume that a union type without null and a field without default value is required
@@ -61,21 +120,23 @@ const requiredAttributesMapping = (fieldDefinition, parentJsonSchema, haveDefaul
     parentJsonSchema.required = parentJsonSchema.required || [];
     parentJsonSchema.required.push(fieldDefinition.name);
   }
-};
+}
 
-function extractNonNullableTypeIfNeeded(typeInput, jsonSchemaInput) {
+function extractNonNullableTypeIfNeeded(typeInput: any, jsonSchemaInput: SpecTypesV2.AsyncAPISchemaDefinition): { type: string, jsonSchema: SpecTypesV2.AsyncAPISchemaDefinition } {
   let type = typeInput;
   let jsonSchema = jsonSchemaInput;
   // Map example to first non-null type
   if (Array.isArray(typeInput) && typeInput.length > 0) {
     const pickSecondType = typeInput.length > 1 && typeInput[0] === 'null';
     type = typeInput[+pickSecondType];
-    jsonSchema = jsonSchema.oneOf[0];
+    if (jsonSchema.oneOf !== undefined) {
+      jsonSchema = jsonSchema.oneOf[0] as SpecTypesV2.AsyncAPISchemaDefinition;
+    }
   }
-  return {type, jsonSchema};
+  return { type, jsonSchema };
 }
 
-const exampleAttributeMapping = (type, example, jsonSchema) => {
+function exampleAttributeMapping(type: any, example: any, jsonSchema: SpecTypesV2.AsyncAPISchemaDefinition): void {
   if (example === undefined || jsonSchema.examples || Array.isArray(type)) return;
 
   switch (type) {
@@ -88,16 +149,16 @@ const exampleAttributeMapping = (type, example, jsonSchema) => {
   default:
     jsonSchema.examples = [example];
   }
-};
+}
 
-const additionalAttributesMapping = (typeInput, avroDefinition, jsonSchemaInput) => {
+function additionalAttributesMapping(typeInput: any, avroDefinition: AvroSchema, jsonSchemaInput: SpecTypesV2.AsyncAPISchemaDefinition): void {
   const __ret = extractNonNullableTypeIfNeeded(typeInput, jsonSchemaInput);
   const type = __ret.type;
   const jsonSchema = __ret.jsonSchema;
 
   exampleAttributeMapping(type, avroDefinition.example, jsonSchema);
 
-  function setAdditionalAttribute(...names) {
+  function setAdditionalAttribute(...names: string[]) {
     names.forEach(name => {
       let isValueCoherent = true;
       if (name === 'minLength' || name === 'maxLength') {
@@ -117,7 +178,9 @@ const additionalAttributesMapping = (typeInput, avroDefinition, jsonSchemaInput)
     setAdditionalAttribute('minimum', 'maximum', 'exclusiveMinimum', 'exclusiveMaximum', 'multipleOf');
     break;
   case 'string':
-    jsonSchema.format = avroDefinition.logicalType;
+    if (avroDefinition.logicalType) {
+      jsonSchema.format = avroDefinition.logicalType;
+    }
     setAdditionalAttribute('pattern', 'minLength', 'maxLength');
     break;
   case 'array':
@@ -126,9 +189,9 @@ const additionalAttributesMapping = (typeInput, avroDefinition, jsonSchemaInput)
   default:
     break;
   }
-};
+}
 
-function validateAvroSchema(avroDefinition) {
+function validateAvroSchema(avroDefinition: AvroSchema): void | never {
   // don't need to use the output from parsing the
   //  avro definition - we're just using this as a
   //  validator as this will throw an exception if
@@ -144,14 +207,14 @@ function validateAvroSchema(avroDefinition) {
  * @param key String | Undefined - the fully qualified name of an avro record
  * @param value JsonSchema - The json schema from the avro record
  */
-function cacheAvroRecordDef(cache, key, value) {
-  if (key) {
+function cacheAvroRecordDef(cache: { [key:string]: AsyncAPISchema }, key: string, value: AsyncAPISchema): void {
+  if (key !== undefined) {
     cache[key] = value;
   }
 }
 
-async function convertAvroToJsonSchema(avroDefinition, isTopLevel, recordCache = {}) {
-  let jsonSchema = {};
+async function convertAvroToJsonSchema(avroDefinition: AvroSchema , isTopLevel: boolean, recordCache: Map<string, SpecTypesV2.AsyncAPISchemaDefinition> | any = {}): Promise<SpecTypesV2.AsyncAPISchemaDefinition> {
+  let jsonSchema: SpecTypesV2.AsyncAPISchemaDefinition = {};
   const isUnion = Array.isArray(avroDefinition);
 
   if (isUnion) {
@@ -164,46 +227,56 @@ async function convertAvroToJsonSchema(avroDefinition, isTopLevel, recordCache =
   jsonSchema.type = typeMappings[type];
 
   switch (type) {
-  case 'int':
+  case 'int': {
     jsonSchema.minimum = INT_MIN;
     jsonSchema.maximum = INT_MAX;
     break;
-  case 'long':
+  }
+  case 'long': {
     jsonSchema.minimum = LONG_MIN;
     jsonSchema.maximum = LONG_MAX;
     break;
-  case 'bytes':
+  }
+  case 'bytes': {
     jsonSchema.pattern = BYTES_PATTERN;
     break;
-  case 'fixed':
+  }
+  case 'fixed': {
     jsonSchema.pattern = BYTES_PATTERN;
     jsonSchema.minLength = avroDefinition.size;
     jsonSchema.maxLength = avroDefinition.size;
     break;
-  case 'map':
+  }
+  case 'map': {
     jsonSchema.additionalProperties = await convertAvroToJsonSchema(avroDefinition.values, false);
     break;
-  case 'array':
+  }
+  case 'array': {
     jsonSchema.items = await convertAvroToJsonSchema(avroDefinition.items, false);
     break;
-  case 'enum':
+  }
+  case 'enum': {
     jsonSchema.enum = avroDefinition.symbols;
     break;
-  case 'float':   // float and double must support the format attribute from the avro type
-  case 'double':
+  }
+  case 'float': // float and double must support the format attribute from the avro type
+  case 'double': {
     jsonSchema.format = type;
     break;
-  case 'record':
+  }
+  case 'record': {
     const propsMap = await processRecordSchema(avroDefinition, recordCache, jsonSchema);
     cacheAvroRecordDef(recordCache, getFullyQualifiedName(avroDefinition), propsMap);
     jsonSchema.properties = Object.fromEntries(propsMap.entries());
     break;
-  default:
-    const cachedRecord = recordCache[avroDefinition];
+  }
+  default: {
+    const cachedRecord = recordCache[getFullyQualifiedName(avroDefinition)];
     if (cachedRecord) {
-      jsonSchema= cachedRecord;
+      jsonSchema = cachedRecord;
     }
     break;
+  }
   }
 
   commonAttributesMapping(avroDefinition, jsonSchema, recordCache);
@@ -221,8 +294,8 @@ async function convertAvroToJsonSchema(avroDefinition, isTopLevel, recordCache =
  * @param jsonSchema the schema for the record.
  * @returns {Promise<Map<string, any>>}
  */
-async function processRecordSchema(avroDefinition, recordCache, jsonSchema) {
-  const propsMap = new Map();
+async function processRecordSchema(avroDefinition: AvroSchema, recordCache: Record<string, SpecTypesV2.AsyncAPISchemaDefinition>, jsonSchema: SpecTypesV2.AsyncAPISchemaDefinition): Promise<Map<string, any>> {
+  const propsMap = new Map<string, any>();
   for (const field of avroDefinition.fields) {
     // If the type is a sub schema it will have been stored in the cache.
     if (recordCache[field.type]) {
@@ -251,12 +324,12 @@ async function processRecordSchema(avroDefinition, recordCache, jsonSchema) {
  * @param avroDefinition the avro schema to be processed
  * @param isTopLevel is this the top level of the schema or is this a sub schema
  * @param recordCache the cache of previously processed record types
- * @returns {Promise<JsonSchema>} the mutated jsonSchema that was provided to the function
+ * @returns {Promise<AsyncAPISchema>} the mutated jsonSchema that was provided to the function
  */
-async function processUnionSchema(jsonSchema, avroDefinition, isTopLevel, recordCache) {
+async function processUnionSchema(jsonSchema: SpecTypesV2.AsyncAPISchemaDefinition, avroDefinition: AvroSchema, isTopLevel: boolean, recordCache: Record<string, SpecTypesV2.AsyncAPISchemaDefinition>): Promise<SpecTypesV2.AsyncAPISchemaDefinition> {
   jsonSchema.oneOf = [];
   let nullDef = null;
-  for (const avroDef of avroDefinition) {
+  for (const avroDef of avroDefinition as any) {
     const def = await convertAvroToJsonSchema(avroDef, isTopLevel, recordCache);
     // avroDef can be { type: 'int', default: 1 } and this is why avroDef.type has priority here
     const defType = avroDef.type || avroDef;
@@ -274,7 +347,6 @@ async function processUnionSchema(jsonSchema, avroDefinition, isTopLevel, record
   return jsonSchema;
 }
 
-module.exports.avroToJsonSchema = async function avroToJsonSchema(avroDefinition) {
-  validateAvroSchema(avroDefinition);
+export async function avroToJsonSchema(avroDefinition: AvroSchema) {
   return convertAvroToJsonSchema(avroDefinition, true);
-};
+}
