+++ test/asyncapi-avro-1.8.2.json
@@ -1,0 +1,22 @@
+{
+  "schemaFormat": "application/vnd.apache.avro;version=1.8.2",
+  "payload": {
+	  "type": "record",
+	  "fields": [
+	    {"name": "name", "type": "string", "example": "Donkey"},
+	    {"name": "age", "type": ["null", "int"], "default": null},
+	    {
+	      "name": "favoriteProgrammingLanguage",
+	      "type": {"name": "ProgrammingLanguage", "type": "enum", "symbols": ["JS", "Java", "Go", "Rust", "C"]}
+	    },
+	    {
+	      "name": "address",
+	      "type": {
+	        "name": "Address",
+	        "type": "record",
+	        "fields": [{"name": "zipcode", "type": "int", "example": 53003}]
+	      }
+	    }
+	  ]
+  }
+}
