+++ test/asyncapi-avro-113-1.9.0.json
@@ -1,0 +1,39 @@
+{
+  "schemaFormat": "application/vnd.apache.avro;version=1.9.0",
+  "payload": [
+    {
+      "type": "record",
+      "name": "Address",
+      "namespace": "com.example",
+      "fields": [
+        {
+          "name": "streetaddress",
+          "type": "string"
+        },
+        {
+          "name": "city",
+          "type": "string"
+        }
+      ]
+    },
+    {
+      "type": "record",
+      "name": "Person",
+      "namespace": "com.example",
+      "fields": [
+        {
+          "name": "firstname",
+          "type": "string"
+        },
+        {
+          "name": "lastname",
+          "type": "string"
+        },
+        {
+          "name": "address",
+          "type": "com.example.Address"
+        }
+      ]
+    }
+  ]  
+}
