+++ test/asyncapi-avro-invalid.json
@@ -1,0 +1,6 @@
+{
+  "schemaFormat": "application/vnd.apache.avro;version=1.9.0",
+  "payload": {
+    "isThisAnAvroSchema": "no"
+  }  
+}
