--- tests/asyncapi-avro-1.9.0-bindings.yaml
@@ -1,19 +1,0 @@
-asyncapi: 2.0.0
-info:
-  title: My API
-  version: '1.0.0'
-channels:
-  mychannel:
-    publish:
-      message:
-        schemaFormat: application/vnd.apache.avro;version=1.9.0
-        payload:
-          $ref: 'schemas/Person-1.9.0-namespace.avsc'
-        bindings:
-          kafka:
-            key:
-              $ref: 'schemas/Person-1.9.0-namespace.avsc'
-          mqtt:
-            x-test:
-              type: string
-
