--- tests/asyncapi-avro-113-1.9.0.yaml
@@ -1,11 +1,0 @@
-asyncapi: 2.0.0
-info:
-  title: My API
-  version: '1.0.0'
-channels:
-  mychannel:
-    publish:
-      message:
-        schemaFormat: application/vnd.apache.avro;version=1.9.0
-        payload:
-          $ref: 'schemas/issue-113-testcase.avsc'
