--- tests/asyncapi-avro-148-record-references.yml
@@ -1,18 +1,0 @@
-asyncapi: 2.4.0
-info:
-  title: Validate record references generation
-  version: '1.0.0'
-
-channels:
-  test-topic:
-    description: "Test"
-
-components:
-  messages:
-    ValidateRecordReferences:
-      name: ValidateRecordReferences
-      title: ValidateRecordReferences
-      schemaFormat: 'application/vnd.apache.avro;version=1.9.0'
-      payload:
-        $ref: 'schemas/issue-148-testcase.avsc'
-
