--- tests/schemas/Invalid.avsc
@@ -1,3 +1,0 @@
-{
-  "isThisAnAvroSchema": "no"
-}
