--- tests/schemas/Person-1.9.0-namespace.avsc
@@ -1,22 +1,0 @@
-{
-  "name": "Person",
-  "namespace": "com.company",
-  "type": "record",
-  "fields": [
-    {"name": "name", "type": "string", "example": "Donkey"},
-    {"name": "age", "type": ["null", "int"], "default": null, "example": 123},
-    {
-      "name": "favoriteProgrammingLanguage",
-      "type": {"name": "ProgrammingLanguage", "type": "enum", "symbols": ["JS", "Java", "Go", "Rust", "C"], "default": "JS"}
-    },
-    {
-      "name": "address",
-      "type": {
-        "name": "Address",
-        "type": "record",
-        "fields": [{"name": "zipcode", "type": "int", "example": 53003}]
-      }
-    },
-    {"name": "someid", "type": "string", "logicalType": "uuid"}
-  ]
-}
