--- tests/schemas/issue-148-testcase.avsc
@@ -1,64 +1,0 @@
-{
-   "type":"record",
-   "name":"RecordWithReferences",
-   "namespace":"com.example",
-   "fields":[
-      {
-         "name":"Record1",
-         "type":{
-            "type":"record",
-            "name":"Record1",
-            "doc":"Reused in other fields",
-            "fields":[
-               {
-                  "name":"string",
-                  "type":"string",
-                  "doc":"field in Record1"
-               }
-            ]
-         }
-      },
-      {
-         "name":"FieldThatDefineRecordInUnion",
-         "type":[
-            "null",
-            {
-               "type":"record",
-               "name":"RecordDefinedInUnion",
-               "namespace":"com.example.model",
-               "doc":"",
-               "fields":[
-                  {
-                     "name":"number",
-                     "type":"long",
-                     "doc":"field in RecordDefinedInUnion",
-                     "minimum": 0,
-                     "maximum": 2
-                  }
-               ]
-            }
-         ],
-         "default":null
-      },
-      {
-         "name":"FieldThatReuseRecordDefinedInUnion",
-         "type":[
-            "null",
-            "com.example.model.RecordDefinedInUnion"
-         ],
-         "default":null
-      },
-      {
-         "name":"FieldThatReuseRecord1",
-         "type":[
-            "null",
-            "Record1"
-         ],
-         "default":null
-      },
-      {
-        "name": "simpleField",
-        "type": "string"
-      }
-   ]
-}
