+++ tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+      "outDir": "./esm",
+      "baseUrl": "./src",
+      "target": "es6",
+      "module": "esnext",
+      "lib": [
+        "esnext"
+      ],
+      "declaration": true,
+      "allowJs": true,
+      "skipLibCheck": true,
+      "esModuleInterop": true,
+      "allowSyntheticDefaultImports": true,
+      "strict": true,
+      "forceConsistentCasingInFileNames": true,
+      "noFallthroughCasesInSwitch": true,
+      "moduleResolution": "node",
+      "resolveJsonModule": true,
+      "isolatedModules": true,
+    },
+    "include": [
+      "src"
+    ]
+  }
